Let pinchhandler operate on QQuickItem properties
This requires coordinate system mapping that varies with the
transformOrigin.

The properties exposed from PinchHandler (rotation, scale and translation)
are currently relative to the point when the pinch became active.
(Therefore, rotation, will reset back to 0 when a new pinch is activated).

Its still unclear how the properties that limits the transform should
influence. With this patch, they are like this:

* {min,max}imumRotation applies to the actual rotation of the item.
* {min,max}imumScale applies to the actual scale of the item.
* {min,max}imum{X,Y} applies to the actual position of the item. (This has
  some unfortunate side-effects when the item is scaled or rotated, since
  the items actual position will change as it rotates)

In addition, the behavior described above means that the limits won't
have any effect if there is no target item.

Change-Id: I279fb03667cd75324e8337039ae2594658265d13
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>