Doc: SwipeDelegate: document grouped signals with \qmlsignal
This makes them appear in the list of signals at the top of
the documentation. The same was done for methods earlier.

a506a0ac added the swipe.opened() and swipe.closed() signals
but missed a changelog entry:

[ChangeLog][Controls][SwipeDelegate] Added swipe.opened() and
swipe.closed() signals, which are emitted when the delegate has
been opened or closed by swipe, and the respective transition has
finished.

Change-Id: Iacd2befc691ab7ed8209736595d8ad98e86e637a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>