SwipeDelegate: rename swipe.rebound to swipe.transition
The "rebound" name originates from Flickable::rebound, but for swipe
delegate the transition is not only applied on release, but also when
calling swipe.open() or swipe.close(). Therefore the "rebound" name
feels a bit off in this context.

Furthermore, this patch adds the missing docs and a changelog entry.

[ChangeLog][Controls][SwipeDelegate] Added a swipe.transition
property that holds the transition that is applied when a swipe
is released, or swipe.open() or swipe.close() is called.

Change-Id: Ic38ec850c64dd21b8d9deb08609172c6cb0f6d71
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>