build a vector of child-filtering parents before delivery of pointer event
Formerly during normal mouse or touch event delivery, sending it to
the Item needed to be done via QQuickWindow::sendEvent, which would
then call sendFilteredMouseEvent, which is a recursive function to
visit all the item's parents, check whether filtersChildMouseEvents()
returns true, if so then return early if childMouseEventFilter()
returns true.  This is the mechanism by which Flickable (for example)
can monitor the movements of an eventpoint even while one of its
children has an exclusive grab, and can steal the grab away.
Now, we do this by building a vector of such parents first, then
visiting them in order.  It might be more efficient, it eliminates
the recursion, and should eliminate the need for a QSet to ensure that
we don't visit the same parent more than once.  We can't change the
behavior of QQuickWindow::sendEvent() because it's public API, but
now we don't use it as much internally.

Change-Id: I686fc5612c66eac09ec05c381a648ec65dec3923
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>