qml: use QUrl::fromUserInput("file.qml", QDir::currentPath())
After change dc6b73390b262b9554599cbf40539763b1280261 in qtbase,
the qml runtime was failing at an assertion in
QQmlTypeLoader::getType() which tries to avoid relative URLs
and relative directory paths.  It's important to convert
relative paths to fully-qualified paths when converting to
an URL, because QQmlTypeLoader uses it as a cache key, and
we want to avoid ambiguity about which file is being cached.

Task-number: QTBUG-57870
Change-Id: Ib984cf722009f5f04cb67fffbc52d12bcc98df89
Reviewed-by: David Faure <david.faure@kdab.com>
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>