Fix a crash with focus handling when destructing QQAppWindow
Because QQuickApplicationWindow connected
  SIGNAL(activeFocusItemChanged())
to
  SLOT(_q_updateActiveFocus()));

it would enter _q_updateActiveFocus() from the dtor of QQuickWindow
(because dtor of QQuickWindow would emit activeFocusItemChanged()).
At that point the QQuickApplicationWindow object for the member function
_q_updateActiveFocus() was already destroyed, so we would crash.

Instead, make sure we clear the focus as early as possible, and then
disconnect in case activeFocusItemChanged() is emitted after
QQuickApplicationWindow is destroyed.

Task-number: QTBUG-57846
Change-Id: I4b1999e647b970394436a2d462b7f352f1c8a811
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>