Clarify JIT selection for 32-bit and 64-bit architectures
Instead of relying on the somewhat "cryptic" ILP32 macro on x86/x86-64,
we can also use the macro that the JIT generated code _really_ relies
on, which is the size of pointers. For x32 those will be 4 despite
Q_PROCESSOR_X86_64, and the JIT will end up disabled. For regular x86
and x86-64 it should get enabled.

The same logic is applied to ARM32/ARM64 to disable the JIT once
machines start showing up with ILP32 data model on ARM64.

Change-Id: Id8dfb203c9b75e8ed5a61dcae28482f8f99be75f
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Dmitry Shachnev <mitya57@gmail.com>