tst_qquickwindow::mouseFiltering: verify that siblings don't filter
A grandchild's parent's sibling (uncle) can be underneath the child
in such a way that it will see the event during normal delivery,
if neither the child nor its parent accepts the event, and yet this
uncle will not get the opportunity to filter because it's not a parent.

Also verify that after an item accepts a press event, subsequent
mouse moves are filtered by the same parent and grandparent, even
if the move has placed the cursor outside their bounds.

Also rename mousePressId to mousePressCount, because the name was
misleading: it's just a counter.

Remove the comment about whether it's a bug that touch events are
filtered: we are constrained by bincompat to use the same virtual
function to filter both kinds of events, for now.

Change-Id: If0fd62730883f3e4e29675580d975395c631e694
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>