QQuickTabBar: fix implicit size calculation
Before, the implicit size of TabBar was calculated based on the content
size of the ListView that TabBar uses internally. The problem was that
ListView calculates the content size based on the explicit size of the
items. There was a circular dependency, because TabBar resized the tabs
to the size of the view. To avoid the circular dependency, TabBar now
calculates the content size based on the total implicit size of the
tabs. As before, explicit size is respected for tabs that have it set.

[ChangeLog][Controls][TabBar] Added contentWidth and contentHeight
properties that are automatically calculated based on the total size
of the tab items, but can be manually overridden if desired. This fixes
an issue that TabBar was not able to reliably calculate an implicit
size, and could in certain scenarios enter an infinite loop due to a
circular dependency between the items' sizes and the tabbar's size.

Task-number: QTBUG-57858
Change-Id: Ie303cbc54247e87b0affc6bf32c7bf99acea4571
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>