Binop assembler cleanup
In preparation for the binop generator to work with different
assemblers, we can bring some of the externally prefixed Assembler::
types into the Binop class namespace. That makes the code easier to read
(less visual noise) and centralizes the dependency to the external type
to one place (the using statements).

Change-Id: Ib7700689196f8e54bba1f3ab6de3bf85ed93ff87
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>