JIT Assembler instruction selection cleanup
In preparation for the isel to work with different assemblers, we can
bring some of the externally prefixed Assembler:: types into the isel
class namespace.  That makes the code easier to read (less visual noise)
and centralizes the dependency to the external type to one place (the
using statements). Similarly the indirect dependency in the
QV4::Assembler sub-class TargetPlatform is replaced with an explicit
JITTargetPlatform alias, so that when QV4::TargetPlatform becomes a
template there is only one place to change in the assembler isel.

Change-Id: Ib36e1ba7298e7be3b344ef027151a744220ab197
Reviewed-by: Lars Knoll <lars.knoll@qt.io>