Editors: fix placeholder text alignment
The internal QQuickPlaceholderText creates an "implicit binding"
to the editor's horizontalAlignment, meaning that the placeholder
respects the editor's horizontal alignment when explicitly set,
and otherwise determines the alignment from the placeholder text's
visual direction.

[ChangeLog][Controls][TextField] Fixed the horizontal alignment
of the placeholder text in right-to-left UIs.

[ChangeLog][Controls][TextArea] Fixed the horizontal alignment
of the placeholder text in right-to-left UIs.

Task-number: QTBUG-55999
Change-Id: If1a8596c35c1920874996277520f1c54430c5f69
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>