Clean up pathitem example grid view
Combine stroke-only and fill-only. Drop the Repeater since it shows
the bad practice of unnecessarily creating multiple PathItem instances.
Drop the PathSvg example for now since this is a discouraged element.
Move the tiger into a separate file and use an async Loader in order to
avoid the 2 step activation.

Change-Id: Ie65eb0284bdb8957f8fd5af7557542a6d044ef61
Reviewed-by: Andy Nichols <andy.nichols@qt.io>