QSGThreadedRenderLoop: Correctly stop the animation timer ...
... when destroying the last window while it is still visible.

For the case of windowDestroyed(), the QQuickWindow can still be
isVisible true and isExposed true. Given these circumstances, we would
not stop the timer (assuming that we needed to keep it running), but
this is actually wrong when we are evaluating a timer change depending
on the window being destroyed.

By calling startOrStopAnimationTimer again, after the window has been
removed from m_windows, we will successfully stop the timer -- the only
cost being an unnecessary polish event being sent from the first call
(which will be ignored).

Task-number: QTBUG-58091
Change-Id: I043ce5d88eafee15019e1c8d41c06d531d1da059
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>