Doc: Improve introduction to Qt Quick SceneGraph Renderer
Mention that the new backends are introduced with Qt 5.8. Also
remove some details that rather belong in subsections, in particular
the mentioning that the Qt Quick API's originally were designed
for OpenGL. It is unclear what consequences this has for the user.
Limitations on backends should be rather spelled out in the
respective subsections.

Change-Id: I21239ea99b89b4c7771e3d79a38c6aa39007d0ea
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>