Fix memory leak in V4
Transitions contain both an id and a set of flags, but the sorting
failed to take the flags into account in the operator<. As a result
we would some times end up with duplicate entries if the same id
was added multiple times with different flags.

If the same id was added again and again with varying flags, this
could lead to an ever expanding list filled with duplicate entries.

Fix this by also taking flags into account in operator< so that
operator< and operator== are symetric and the list gets correctly
sorted.

Change-Id: I6d55c67083e4a09e3eb2952edbec302389421f33
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>
(cherry picked from commit 94324a4ea2261940d1d55dec141c885bf4f57832)