corrections: copy some values from QQuickEventPoint to properties
commit 5c6245b7ff7a941c0a2425f32909c0dc0982e199 in its final form was
not what we need for weak/passive grabs, although earlier versions were.
Only the PointerHandler subclass can decide whether to do an exclusive grab,
and we must always ensure that QQuickPointerSingleHandler::handleEventPoint
is called, even when a point is released, otherwise some types of handlers
aren't possible.

Also, DragHandler needs to retain terminal velocity until after release.
onActiveChanged is a handy place to check the velocity at the point of
release, to start an animation for example.  So when the point is
released, we'd like the velocity property to continue providing the
last-known velocity for as long as possible.  Setting it back to zero
and emitting velocityChanged should be one of the last things it does.
This makes flicking possible.

Change-Id: Ibf3bec27162b016ba3f93269d90959c9ccb74d77
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>