Allow more expressions as base/index in IR subscripts
Arguments and locals can be used directly as the base of a subscript,
so they do not need to be assigned to a temporary first. For indices,
arguments, locals, and constants, can be used just fine for subscripts.

This reduces a whole lot of assignments to temporaries when generating
the IR. Although they will be removed in an optimization pass, they do
consume memory for no good reason, and make SSA transformation take more
time.

Change-Id: Ie2af65b66fecee3e140228a9532c9fab08474f5b
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>