QQuickMouseArea: only start pressAndHoldTimer for accepted events
Most likely because of changes done to mouse event handling in 5.8,
a QQuickMouseArea can now get a mousePressEvent call without a
subsequent mouseReleaseEvent call if the press event was rejected.

A regression seen from this is that QQuickMouseArea starts a
pressAndHold timer upon receiving mouse press, which will never be
'cancelled' since the following mouse release is sent elsewhere.

As a result, TextArea in QtQuickControls will e.g receive a
pressAndHold signal each time the user touches the control, which
again will cause weird behavior like selecting words and
messing with focus on iOS.

This patch will instead change this so that we only start the pressAndHold
timer if QQuickMouseArea accepts the mouse press event.

Change-Id: I43e2f02e2b4f75e7b6761e326e3fec54ea97683a
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>