fix test_mouseDrag on RHEL 7.2
mouseDrag(item, x, y, dx, dy, button, modifiers, delay)
in TestCase.qml adds intermediate mouseMove events if dx or dy
is too large (specifically if Math.round(dx/3) > dragThreshold)
and that is affecting the outcome of this test.  So we need to
stay under that threshold.  The original point of this test is from
f52227f66a7af5692140ad036c06857cd2e7abcf / QTBUG-30188 :
when you drag past the dragThreshold, and the dragged Item "breaks loose"
and starts dragging, the initial distance that it jumps should be
equal to the dragThreshold, not greater.  To test it, we need
to move a distance which is greater than the dragThreshold in
one move event, not break it up into smaller movements.

This reverts commit 432eb3344b0581ea1915840365b71376388a75b2.

Task-number: QTBUG-58025
Change-Id: Ib6da69fb465b58fde1bf8ba56c4c3a1cb584f698
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Liang Qi <liang.qi@qt.io>